Added Pay method

+++ HouseholdIncomeAndExpensesWebbApp/Controllers/BillController.cs
@@ -27,26 +27,21 @@ namespace HouseholdBudgetingApp.Controllers
         [HttpGet]
         public async Task<IActionResult> Index()
         {
-            var userId = User.Id();
-            var model = await billService.AllBillsAsync(userId);
-
+            var model = await billService.AllCurentMonthBillsAsync(User.Id(), billService.GetDate());
+            foreach (var item in model) { item.Payers = await householdService.GetHouseholdMembersAsync(User.Id()); }
+            ViewBag.Date = billService.GetFormatedDate();
             return View(model);
         }
 
         [HttpGet]
         public async Task<IActionResult> Add()
         {
-            var userId = User.Id();
-            DateTime date = DateTime.Now;
-            DateTime monthYearOnly = new DateTime(date.Year, date.Month, 1);
-
-            string formattedDate = monthYearOnly.ToString("MMMM yyyy");
-            ViewBag.Date = formattedDate;
+                        
+            ViewBag.Date = billService.GetFormatedDate();
             var model = new BillFormModel()
             {
-                BillTypes = await billService.GetBillTypesAsync(userId),
-                Payers=await householdService.GetHouseholdMembersAsync(userId),
-                Date=monthYearOnly,
+                BillTypes = await billService.GetBillTypesAsync(User.Id()),
+                Payers=await householdService.GetHouseholdMembersAsync(User.Id()),                
             };
             
             return View(model);
@@ -55,34 +50,32 @@ namespace HouseholdBudgetingApp.Controllers
         [HttpPost]
         public async Task<IActionResult>Add(BillFormModel model)
         {
-            var userId=User.Id();
-            if (!(await billService.GetBillTypesAsync(userId)).Any(b => b.Id == model.BillTypeId))
+            if (string.IsNullOrEmpty(Request.Form["PayerId"]))
             {
-                ModelState.AddModelError(nameof(model.BillTypeId), "Type does not exist.");
+                model.IsPayed = false;
+                model.PayerId = null;
+            }
+            else
+            {
+                model.IsPayed = true;
+                model.PayerId = int.Parse(Request.Form["PayerId"]);
             }
 
-            //DateTime date = DateTime.Now;
+            if (!(await billService.GetBillTypesAsync(User.Id())).Any(b => b.Id == model.BillTypeId))
+            {
+                ModelState.AddModelError(nameof(model.BillTypeId), "Bill Type does not exist.");
+            }
             
-
-            //if (!DateTime.TryParseExact(
-            //    model.Date,"yyyy-MM-dd H:mm",
-            //    CultureInfo.InvariantCulture,
-            //    DateTimeStyles.None,
-            //    out date))
-            //{
-            //    ModelState
-            //        .AddModelError(nameof(model.Date), $"Invalid date! Format must be: yyyy-MM-dd H:mm");
-            //}
 
             if (!ModelState.IsValid)
             {
-                model.BillTypes = await billService.GetBillTypesAsync(userId);
-                model.Payers = await householdService.GetHouseholdMembersAsync(userId);
+                model.BillTypes = await billService.GetBillTypesAsync(User.Id());
+                model.Payers = await householdService.GetHouseholdMembersAsync(User.Id());
 
                 return View(model);
             }
 
-            await billService.CreateBillAsync(model, userId);
+            await billService.CreateBillAsync(model, User.Id());
             return RedirectToAction("Index");
 
         }
@@ -90,39 +83,51 @@ namespace HouseholdBudgetingApp.Controllers
         [HttpGet]
         public async Task<IActionResult> Edit(int id)
         {
-            var userId = User.Id();
+           
+            if (await billService.BillExistsAsync(id) == false)
+            {
+                return BadRequest();
+            }
+            if (await billService.BillIsPayedAsync(id))
+            {
+                return BadRequest();
+            }
+
+            if (await billService.BillBelongsToUserAsync(id, User.Id()) == false)
+            {
+                return Unauthorized();
+            }
+
             var model=await billService.FindBillByIdAsync(id);
-            model.BillTypes = await billService.GetBillTypesAsync(userId);
-            model.Payers = await householdService.GetHouseholdMembersAsync(userId);
+            model.BillTypes = await billService.GetBillTypesAsync(User.Id());            
             return View(model);
         }
         [HttpPost]
         public async Task<IActionResult> Edit(BillFormModel model,int id)
         {
-            var userId = User.Id();
-            if (!(await billService.GetBillTypesAsync(userId)).Any(b => b.Id == model.BillTypeId))
+            if (await billService.BillExistsAsync(id) == false)
             {
-                ModelState.AddModelError(nameof(model.BillTypeId), "Type does not exist.");
+                return BadRequest();
+            }
+            if (await billService.BillIsPayedAsync(id))
+            {
+                return BadRequest();
             }
 
-            //DateTime date = DateTime.Now;
-
-
-            //if (!DateTime.TryParseExact(
-            //    model.Date, "yyyy-MM-dd H:mm",
-            //    CultureInfo.InvariantCulture,
-            //    DateTimeStyles.None,
-            //    out date))
-            //{
-            //    ModelState
-            //        .AddModelError(nameof(model.Date), $"Invalid date! Format must be: yyyy-MM-dd H:mm");
-            //}
+            if (await billService.BillBelongsToUserAsync(id, User.Id()) == false)
+            {
+                return Unauthorized();
+            }
+            if (!(await billService.GetBillTypesAsync(User.Id())).Any(b => b.Id == model.BillTypeId))
+            {
+                ModelState.AddModelError(nameof(model.BillTypeId), "Bill Type does not exist.");
+            }
+            
 
             if (!ModelState.IsValid)
             {
-                model.BillTypes = await billService.GetBillTypesAsync(userId);
-                model.Payers = await householdService.GetHouseholdMembersAsync(userId);
-
+                model.BillTypes = await billService.GetBillTypesAsync(User.Id());
+               
                 return View(model);
             }
 
@@ -130,10 +135,48 @@ namespace HouseholdBudgetingApp.Controllers
             return RedirectToAction("Index");
         }
 
+        [HttpPost]
+        public async Task<IActionResult> Pay(BillViewModel model, int id)
+        {
+            if (await billService.BillExistsAsync(id) == false)
+            {
+                return BadRequest();
+            }
+
+            if (await billService.BillBelongsToUserAsync(id, User.Id()) == false)
+            {
+                return Unauthorized();
+            }
+           
+
+            if (!(await householdService.GetHouseholdMembersAsync(User.Id())).Any(m => m.Id == model.PayerId))
+            {
+                ModelState.AddModelError(nameof(model.PayerId), "Household Member does not exist.");
+            }
+            if (!ModelState.IsValid)
+            {
+                return RedirectToAction("Index");
+            }
+
+
+
+            await billService.PayBillAsync(model, id);
+            return RedirectToAction("Index");
+        }
+
 
         [HttpPost]
         public async Task<IActionResult> Delete(int id)
         {
+            if (await billService.BillExistsAsync(id) == false)
+            {
+                return BadRequest();
+            }
+
+            if (await billService.BillBelongsToUserAsync(id, User.Id()) == false)
+            {
+                return Unauthorized();
+            }
             await billService.DeleteBillByIdAsync(id);
 
             return RedirectToAction("Index");
